+++ CHANGELOG.md
@@ -38,6 +38,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add automatic highlighting of bars on hover to bar series.
 - Add labelHighlight function.
 - Add labelFind function find labels by key.
+- Add labelFindByIndex function.
+- Add labelFindByFilter function.
 - Add axisTickHighlight function.
 - Add axisTickFindByIndex function.
 - Add chartBarHoverBar function to apply hover over bar effects to bar chart.